Various fixes * Fixed warnings about signed/unsigned mismatch by changing to unsigned a few values that are only ever assigned or compared to unsigned values, yet were signed for some reason. * Fixed TextEntryPanel's event handler missing the check about whether it was the active panel. Among other annoyances, it made impossible to do a search & replace operation on a different tab than the first. * Added some commented-out stuff about the Strife demo conversation script format. It's more for documentation than anything else, which is why it's commented out.

+++ src/Archive/ArchiveEntry.h
@@ -117,7 +117,7 @@ public:
 	int		getTypeReliability() { return (type ? (getType()->getReliability() * reliability / 255) : 0); }
 	bool	isInNamespace(string ns);
 
-	int index_guess; // for speed
+	size_t	index_guess; // for speed
 };
 
 #endif//__ARCHIVEENTRY_H__
